Move useIframeDocument around
